requirements.txt
allure-pytest
allure-python-commons
pytest==7.1.1
selenium==4.1.3
webdriver-manager==3.4.0